pyproject.toml
[project]
name = "flatshop"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jack", email = ""},
]
dependencies = [
    "telebot>=0.0.5",
    "gkeepapi>=0.16.0",
    "bs4>=0.0.2",
    "python-dotenv>=1.0.1",
    "ollama>=0.2.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
